fixing office user assign , craeting ambar module , adding config needed in services

+++ Modules/Hr/Http/Controllers/Employee/ContractController.php
@@ -2,19 +2,21 @@
 
 namespace Modules\Hr\Http\Controllers\Employee;
 
+use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Support\Facades\DB;
 use Modules\Hr\Entities\Employee\Contract;
 use Modules\Hr\Entities\Employee\Employee;
 use App\Traits\ApiResponse;
-use App\Traits\DocumentUploader;
 use App\Traits\Query;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
+use Modules\Hr\Traits\DocumentUploader;
 
 class ContractController extends Controller
 {
 
-    use ApiResponse, Query, DocumentUploader;
+    use ApiResponse, Query, DocumentUploader, ValidatesRequests;
 
     public function index(Request $request)
     {
@@ -26,7 +28,9 @@ class ContractController extends Controller
             'employee_status' => ['sometimes', 'required', 'in:0,1,2'],
         ]);
         try {
-            $contract = Contract::with(['employee:id,human_id', 'employee.human:id,name,surname'])->whereHas('employee', function ($q) use ($request) {
+            $contract = Contract::with([
+                'employee:id,user_id,is_active', 'employee.user:id,name,surname' , 'position:id,name'
+            ])->whereHas('employee', function ($q) use ($request) {
                 $q->where('company_id', $request->get('company_id'));
                 if ($request->has('employee_status') and $request->get('employee_status') != 2) {
                     $q->where('is_active', $request->get('employee_status'));
@@ -34,12 +38,12 @@ class ContractController extends Controller
                     $q->where('is_active', true);
                 }
             });
+
             if ($request->has('status') and $request->get('status') != 2) {
                 $contract->where('is_active', $request->get('status'));
             } else {
                 $contract->where('is_active', true);
             }
-
 
             $contract = $contract->paginate();
 
@@ -59,54 +63,62 @@ class ContractController extends Controller
             'position_id' => ['required', 'integer'],
             'salary' => ['required', 'numeric'],
             'contract' => ['sometimes', 'mimes:pdf,doc,docx'],
-            'from' => ['sometimes', 'required', 'date'],
-            'to' => ['sometimes', 'required', 'date'],
+            'start_date' => ['sometimes', 'required', 'date'],
+            'end_date' => ['sometimes', 'required', 'date'],
             'employee_id' => ['required', 'integer'],
             'company_id' => ['required', 'integer']
         ]);
-        try {
-            \DB::beginTransaction();
-
-            $response = "ok";
-
-            if ($notExists = $this->companyInfo($request->get('company_id'), $request->only([
-                'department_id', 'section_id', 'position_id', 'sector_id'
-            ]))) return $notExists;
-
-            $employee = Employee::where('id', $request->get('employee_id'))
-                ->where('company_id', $request->get('company_id'))
-                ->where('is_active', true)
-                ->first(['id', 'human_id']);
-
-            if (!$employee) return $this->errorResponse(trans('response.employeeNotFound'));
+        DB::beginTransaction();
 
 
-            $data = $request->only([
-                'department_id', 'section_id', 'sector_id', 'position_id', 'salary', 'from', 'to', 'employee_id'
-            ]);
+        if ($notExists = $this->companyInfo($request->get('company_id'), $request->only(['department_id', 'section_id', 'position_id', 'sector_id']))) return $notExists;
+
+        $employee = Employee::where('id', $request->get('employee_id'))
+            ->where('company_id', $request->get('company_id'))
+            ->first(['id', 'is_active']);
+
+        if (!$employee) return $this->errorResponse(trans('response.employeeNotFound'));
+
+        if (!$employee->is_active) return $this->errorResponse(trans('response.employeeIsOut'));
+
+        Contract::create(array_merge($request->only(
+            [
+                'department_id',
+                'section_id',
+                'sector_id',
+                'position_id',
+                'salary',
+                'contract',
+                'start_date',
+                'end_date',
+                'employee_id',
+            ]
+        ), [
+                'employee_id' => $employee->id
+            ]
+        ));
 
 
-            if ($request->has('contract') and $request->hasFile('contract'))
-                $data['contract'] = $this->save($request->file('contract'), $request->get('company_id'), 'contracts');
+        $data = $request->only([
+            'department_id', 'section_id', 'sector_id', 'position_id', 'salary', 'start_date', 'end_date', 'employee_id'
+        ]);
 
 
-            $ctr = Contract::create($data);
+        if ($request->has('contract') and $request->hasFile('contract'))
+            $data['contract'] = $this->save($request->file('contract'), $request->get('company_id'), 'contracts');
 
 
-            Contract::where('employee_id', $request->get('employee_id'))
-                ->where('id', '!=', $ctr->id)
-                ->update(['is_active' => false]);
+        $ctr = Contract::create($data);
 
 
-            if ($request->has('return_human')) $response = ['human_id' => $employee->human_id];
+        Contract::where('employee_id', $request->get('employee_id'))
+            ->where('id', '!=', $ctr->id)
+            ->update(['is_active' => false]);
 
 
-            \DB::commit();
-            return $this->successResponse($response);
-        } catch (\Exception $e) {
-            \DB::rollBack();
-            return $this->errorResponse(trans('response.tryLater'));
-        }
+        DB::commit();
+        return $this->successResponse('ok');
+
     }
 
     public function update(Request $request, $id)
@@ -118,43 +130,47 @@ class ContractController extends Controller
             'position_id' => ['sometimes', 'required', 'integer'],
             'salary' => ['sometimes', 'required', 'numeric'],
             'contract' => ['sometimes', 'mimes:pdf,doc,docx'],
-            'from' => ['sometimes', 'required', 'date'],
-            'to' => ['sometimes', 'required', 'date'],
+            'start_date' => ['sometimes', 'required', 'date'],
+            'end_date' => ['sometimes', 'required', 'date'],
             'employee_id' => ['required', 'integer'],
             'company_id' => ['required', 'integer'],
+            'is_active' => ['sometimes' , 'required' , 'boolean']
         ]);
-        $data = $request->only(['department_id', 'section_id', 'sector_id', 'position_id', 'salary', 'from', 'to', 'contract']);
+        $data = $request->only(['department_id', 'section_id', 'sector_id', 'position_id', 'salary', 'start_date', 'end_date', 'contract' , 'is_active']);
         if (!$data) return $this->errorResponse(trans('response.nothing'));
         try {
-            \DB::beginTransaction();
+            DB::beginTransaction();
 
             if ($notExists = $this->companyInfo($request->get('company_id'), $request->only([
                 'department_id', 'section_id', 'position_id', 'sector_id'
             ]))) return $notExists;
+
             $employee = Employee::where('id', $request->get('employee_id'))
                 ->where('company_id', $request->get('company_id'))
-                ->where('is_active', true)
-                ->first(['id', 'human_id']);
+                ->first(['id', 'is_active']);
+
             if (!$employee) return $this->errorResponse(trans('response.employeeNotFound'), 404);
 
-            $contract = Contract::where('employee_id', $employee->id)->where('id', $id)->first(['position_id']);
+            if (!$employee->is_active) return $this->errorResponse(trans('response.employeeNotActiveWorker'), 400);
 
+            $contract = Contract::where('id', $id)->where('employee_id', $employee->id)->first(['position_id']);
 
-            if (!$contract) return $this->errorResponse([trans('response.contractNotFound'),$employee->id,$id], 404);
+            if (!$contract) return $this->errorResponse(trans('response.contractNotFound'), 404);
+//            if (!$contract->is_active) return $this->errorResponse(trans('response.contractIsNotActive'), 404);
 
-            if ($request->has('contract') and $request->hasFile('contract'))
+            if ($request->hasFile('contract'))
                 $data['contract'] = $this->save($request->file('contract'), $request->get('company_id'), 'contracts');
 
             Contract::where('id', $id)->update($data);
 
-            \DB::commit();
-            return $this->dataResponse([
-                'human' => $employee->human_id,
-                'change_position' => $request->has('position_id') ? $request->get('position_id') != $contract->position_id : false,
-            ]);
+            if ($request->get('is_active')) Contract::where('employee_id' , $employee->id) -> update(['is_active' , false]);
+
+            DB::commit();
+            return $this->successResponse('ok');
         } catch (\Exception $e) {
-            \DB::rollBack();
-            return $this->errorMessage(trans('response.tryLater'));
+            dd($e);
+            DB::rollBack();
+            return $this->errorResponse(trans('response.tryLater'));
         }
     }
 
@@ -163,26 +179,22 @@ class ContractController extends Controller
         $this->validate($request, [
             'company_id' => ['required', 'integer']
         ]);
-        try {
 
-            $employee = Employee::where('id', $request->get('employee_id'))
-                ->where('company_id', $request->get('company_id'))
-                ->first(['id', 'human_id' , 'is_active']);
-            if (!$employee) return $this->errorResponse(trans('response.employeeNotFound'), 404);
-            if (!$employee->is_active) return $this->errorResponse(trans('response.employeeNotActiveWorker'), 400);
+        $employee = Employee::where('id', $request->get('employee_id'))
+            ->where('company_id', $request->get('company_id'))
+            ->first(['id', 'is_active']);
+        if (!$employee) return $this->errorResponse(trans('response.employeeNotFound'), 404);
+        if (!$employee->is_active) return $this->errorResponse(trans('response.employeeNotActiveWorker'), 400);
 
-            $contract = Contract::where('employee_id',$employee->id)->where('id', $id)->first(['id', 'is_active']);
+        $contract = Contract::where('employee_id', $employee->id)->where('id', $id)->first(['id', 'is_active']);
 
-            if (!$contract) return $this->errorResponse(trans('response.contractNotFound'), 404);
+        if (!$contract) return $this->errorResponse(trans('response.contractNotFound'), 404);
 
-//            if ($contract->is_active) return $this->errorResponse(trans('response.cannotDeleteActiveContract'));
+        if ($contract->is_active) return $this->errorResponse(trans('response.cannotDeleteActiveContract'));
 
-            $contract->delete();
+        $contract->delete();
 
-            return $this->successResponse(['is_active' => $contract->is_active , 'human_id' => $employee->human_id]);
-        } catch (\Exception $e) {
-            return $this->errorResponse(trans('response.tryLater'), Response::HTTP_INTERNAL_SERVER_ERROR);
-        }
+        return $this->successResponse();
     }
 
     public function show(Request $request, $id)
@@ -190,14 +202,11 @@ class ContractController extends Controller
         $this->validate($request, [
             'company_id' => ['required', 'integer'],
         ]);
-        try {
-            $contract = Contract::with(['employee', 'employee.human'])->whereHas('employee', function ($q) use ($request) {
-                $q->where('company_id', $request->get('company_id'));
-            })->where('id', $id)->first();
-            return $this->successResponse($contract);
-        } catch (\Exception $e) {
-            return $this->errorResponse(trans('response.tryLater'));
-        }
+        $contract = Contract::with(['position' , 'department' , 'section' , 'sector' , 'employee:id,is_active', 'employee.user:id,name,surname'])
+            ->whereHas('employee', function ($q) use ($request) {
+            $q->where('company_id', $request->get('company_id'));
+        })->where('id', $id)->first();
+        return $this->successResponse($contract);
     }
